+++ tests/core/test_index.py
@@ -3,12 +3,14 @@
 
 import datetime
 import logging
+import pickle
 from itertools import permutations
 
 import numpy as np
 import pandas as pd
 import pyarrow as pa
 import pytest
+import pytz
 from hypothesis import assume, given
 from pandas.testing import assert_series_equal
 
@@ -367,16 +369,26 @@ def test_index_as_flat_series_date():
     assert_series_equal(ser, expected)
 
 
-def test_index_store_roundtrip_ts(store):
+@pytest.mark.parametrize(
+    "dtype, timestamps",
+    [
+        (pa.timestamp("ns"), [pd.Timestamp("2017-01-01"), pd.Timestamp("2017-01-02")]),
+        (
+            pa.timestamp("ns"),
+            [
+                pd.Timestamp("2017-01-01", tzinfo=pytz.timezone("EST")),
+                pd.Timestamp("2017-01-02", tzinfo=pytz.timezone("EST")),
+            ],
+        ),
+    ],
+)
+def test_index_store_roundtrip_ts(store, dtype, timestamps):
     storage_key = "dataset_uuid/some_index.parquet"
     index1 = ExplicitSecondaryIndex(
         column="col",
-        index_dct={
-            pd.Timestamp("2017-01-01"): ["part_1", "part_2"],
-            pd.Timestamp("2017-01-02"): ["part_3"],
-        },
+        index_dct=dict(zip(timestamps, [["part_1", "part_2"], ["part_3"]])),
         index_storage_key=storage_key,
-        dtype=pa.timestamp("ns"),
+        dtype=dtype,
     )
     key1 = index1.store(store, "dataset_uuid")
 
@@ -448,6 +460,18 @@ def test_index_raises_null_dtype():
         ),
         (pa.timestamp("ns"), "2018-01-01", pd.Timestamp("2018-01-01").to_datetime64()),
         (pa.date32(), "2018-01-01", datetime.date(2018, 1, 1)),
+        (
+            pa.timestamp("ns", tz=pytz.timezone("Europe/Berlin")),
+            pd.Timestamp("2018-01-01", tzinfo=pytz.timezone("Europe/Berlin")),
+            pd.Timestamp(
+                "2018-01-01", tzinfo=pytz.timezone("Europe/Berlin")
+            ).to_datetime64(),
+        ),
+        (
+            pa.timestamp("ns", tz=pytz.timezone("Europe/Berlin")),
+            "2018-01-01",  # Naive date, is interpreted as being UTC
+            pd.Timestamp("2018-01-01", tzinfo=pytz.timezone("UTC")).to_datetime64(),
+        ),
     ],
 )
 def test_index_normalize_value(dtype, value, expected):
@@ -619,3 +643,79 @@ def test_index_uint():
         },
     )
     assert index.dtype == "uint64"
+
+
+@pytest.mark.parametrize(
+    "key",
+    [
+        True,  # pa.bool_()
+        1,  # pa.int64()
+        1.1,  # pa.float64()
+        b"x",  # pa.binary()
+        "ö",  # pa.string()
+        pd.Timestamp("2018-01-01").to_datetime64(),  # pa.timestamp("ns")
+        pd.Timestamp(
+            "2018-01-01", tzinfo=pytz.timezone("Europe/Berlin")
+        ).to_datetime64(),  # pa.timestamp("ns")
+        datetime.date(2018, 1, 1),  # pa.date32()
+    ],
+)
+def test_serialization(key):
+    """Check index remains consistent after serializing and de-serializing"""
+    index = ExplicitSecondaryIndex(
+        column="col", index_dct={key: ["part_2", "part_4", "part_1"]}
+    )
+    index2 = pickle.loads(pickle.dumps(index))
+
+    assert index == index2
+
+
+@pytest.mark.parametrize(
+    "key",
+    [
+        True,  # pa.bool_()
+        1,  # pa.int64()
+        1.1,  # pa.float64()
+        b"x",  # pa.binary()
+        "ö",  # pa.string()
+        pd.Timestamp("2018-01-01").to_datetime64(),  # pa.timestamp("ns")
+        pd.Timestamp(
+            "2018-01-01", tzinfo=pytz.timezone("Europe/Berlin")
+        ).to_datetime64(),  # pa.timestamp("ns")
+        datetime.datetime(
+            2018, 1, 1, 12, 30
+        ),  # pa.timestamp("us") (initial) => pa.timestamp("ns") (after loading)
+        datetime.datetime(
+            2018, 1, 1, 12, 30, tzinfo=pytz.timezone("Europe/Berlin")
+        ),  # pa.timestamp("ns")
+        datetime.date(2018, 1, 1),  # pa.date32()
+    ],
+)
+def test_serialization_normalization(key):
+    """
+    Check that index normalizes values consistently after serializing.
+
+    This is helpful to ensure correct behavior for cases such as when
+    key=`datetime.datetime(2018, 1, 1, 12, 30)`, as this would be parsed to
+    `pa.timestamp("us")` during index creation, but stored as `pa.timestamp("ns")`.
+    """
+    index = ExplicitSecondaryIndex(
+        column="col", index_dct={key: ["part_2", "part_4", "part_1"]}
+    )
+    index2 = pickle.loads(pickle.dumps(index))
+
+    assert index.normalize_value(index.dtype, key) == index2.normalize_value(
+        index2.dtype, key
+    )
+
+
+def test_serialization_no_indices(store):
+    index = ExplicitSecondaryIndex(column="col", index_dct={1: ["part_1"]})
+    storage_key = index.store(store=store, dataset_uuid="uuid")
+
+    # Create index without `index_dct`
+    index = ExplicitSecondaryIndex(column="col", index_storage_key=storage_key)
+
+    index2 = pickle.loads(pickle.dumps(index))
+
+    assert index == index2