+++ CHANGES.rst
@@ -10,9 +10,11 @@ Version 3.1.0 (2019-XX-XX)
 - fix ``FutureWarning`` in ``filter_array_like``
 - remove ``funcsigs`` requirement
 - Implement reference ``io.eager`` implementation, adding the functions:
-    + ``io.eager.garbage_collect_dataset``
-    + ``io.eager.index.build_dataset_indices``
-    + ``io.eager.update_dataset_from_dataframes``
+
+    - :meth:`~kartothek.io.eager.garbage_collect_dataset`
+    - :meth:`~kartothek.io.eager.build_dataset_indices`
+    - :meth:`~kartothek.io.eager.update_dataset_from_dataframes`
+
 - fix ``_apply_partition_key_predicates`` ``FutureWarning``
 - serialize :class:`~kartothek.core.index.ExplicitSecondaryIndex` to parquet
 - improve messages for schema violation errors