+++ CHANGES.rst
@@ -18,6 +18,11 @@ Version 3.1.0 (2019-XX-XX)
 - fix ``_apply_partition_key_predicates`` ``FutureWarning``
 - serialize :class:`~kartothek.core.index.ExplicitSecondaryIndex` to parquet
 - improve messages for schema violation errors
+- Ensure binary column names are read as type ``str``:
+
+    - Ensure dataframe columns are of type ``str`` in :func:`~kartothek.core.common_metadata.empty_dataframe_from_schema`
+    - Testing: create :func:`~kartothek.io.testing.read.test_binary_column_metadata` which checks column names stored as
+      ``bytes`` objects are read as type ``str``
 
 **Breaking:**