+++ CHANGES.rst
@@ -29,6 +29,10 @@ Version 3.1.0 (2019-XX-XX)
 
 - fix issue where unreferenced files were not being removed when deleting an entire dataset
 
+- support nested :class:`~kartothek.io_components.metapartition.MetaPartition`
+  in :meth:`~kartothek.io_components.metapartition.MetaPartition.add_metapartition`.
+  This fixes issue https://github.com/JDASoftwareGroup/kartothek/issues/40 .
+
 **Breaking:**
 
 - categorical normalization was moved from :meth:`~kartothek.core.common_metadata.make_meta` to